Fix issue where some VLI systems wouldn't be recognized (#216)
We were using "dmidecode" to determine what type of system we were
running on, with expected answers of "UV300" or "HPE". Turns out
that the system can also report as "HPE Integrity MC990 X Server".

If "MC990 X" is anywhere in the response, allow that as recognition
of a VLI system.